+++ client/src/index.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { Desktop } from "./screens/Desktop.jsx";
+import { Desktop } from "./screens/Desktop/Desktop.jsx";
 import { SignUp } from "./screens/SignUp.jsx";
 import { BrowserRouter, Routes, Route } from "react-router";
 import DashboardPreview from "./screens/DashboardPreview.jsx";
@@ -12,6 +12,7 @@ import { store } from "./store/store.js";
 import React, { useEffect } from "react";
 import { login } from "./store/slices/authSlice";
 import { Settings } from "./screens/Settings.jsx";
+import './style.css';
 
 function App() {
   const dispatch = useDispatch();
@@ -26,7 +27,7 @@ function App() {
 
   return (
     <Routes>
-      <Route path="/" element={<Desktop />} />
+      <Route index element={<Desktop />} />
       <Route path="/signup" element={<SignUp />} />
       <Route path="/login" element={<Login />} />
       <Route path="/dashboard" element={<DashboardPreview />} />
