+++ client/src/screens/Desktop/Desktop.css
@@ -1,47 +1,20 @@
-  body {
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    min-height: 100vh;
-    background: #f0f0f0;
-  }
-  
-  .desktop {
+.desktop {
     background-color: #ffffff;
-    width: 1440px;
+    width: cover;
     min-height: 100vh;
-    overflow: hidden;
-    position: relative;
+    
   }
-  
-  .desktop .overlap {
-    height: 4546px;
-    position: relative;
+
+  html{
+    scroll-behavior: smooth;
   }
   
   .desktop .unsplash {
     height: 1034px;
-    left: 0;
     object-fit: cover;
-    position: absolute;
+    position: relative;
     top: 0;
-    width: 1440px;
-  }
-  
-  .desktop .rectangle {
-    background: linear-gradient(
-      180deg,
-      rgba(102, 102, 102, 0) 0%,
-      rgba(0, 0, 0, 0.5) 50%,
-      rgba(0, 0, 0, 0.75) 75%,
-      rgb(0, 0, 0) 100%
-    );
-    height: 1440px;
-    left: 520px;
-    position: absolute;
-    top: 114px;
-    transform: rotate(-90.00deg);
-    width: 400px;
+    width: 1980px;
   }
   
   .desktop .frame {
@@ -55,7 +28,7 @@
     padding: 20px 80px;
     position: absolute;
     top: 0;
-    width: 1440px;
+    width: 100%;
   }
   
   .desktop .div {
@@ -232,18 +205,19 @@
     margin-top: -1.00px;
     width: 900px;
     position: relative;
-    text-shadow: 2px 5px rgba(0,0,0,0.2);
+    text-shadow: 2px 5px 3px rgba(0,0,0,0.3);
   }
   
   .desktop .p {
     color: #ffffff;
     font-family: "DM Sans", Helvetica;
-    font-size: 16px;
+    font-size: 20px;
     font-weight: 400;
     letter-spacing: 0;
     line-height: normal;
     position: relative;
     width: 718px;
+    text-shadow: 2px 2px 5px rgba(0,0,0,1);
   }
   
   .desktop .button {
@@ -285,7 +259,7 @@
     left: 0;
     position: absolute;
     top: 1034px;
-    width: 1440px;
+    width: 100%;
   }
   
   .desktop .metrics-section {
@@ -355,6 +329,8 @@
     letter-spacing: var(--text-xl-normal-letter-spacing);
     line-height: var(--text-xl-normal-line-height);
     position: relative;
+    left: 50px;
+    top: -20px;
   }
   
   .desktop .container {
@@ -474,7 +450,7 @@
     flex: 0 0 auto;
     flex-direction: column;
     gap: 42px;
-    padding: 120px;
+    padding: 100px;
     position: relative;
     width: 100%;
   }
@@ -750,9 +726,9 @@
   
   .desktop .overlap-group {
     height: 488px;
-    left: -409px;
+    left: -390px;
     position: relative;
-    width: 2046px;
+    width: 100%;
   }
   
   .desktop .container-2 {
@@ -762,7 +738,7 @@
     left: 0;
     position: absolute;
     top: 0;
-    width: 1716px;
+    width: 2200px;
   }
   
   .desktop .card-2 {
@@ -987,10 +963,9 @@
       rgb(255, 255, 255) 100%
     );
     height: 483px;
-    left: 330px;
     position: absolute;
     top: 0;
-    width: 1440px;
+    width: 100%;
   }
   
   .desktop .frame-13 {
@@ -1003,7 +978,7 @@
     justify-content: center;
     padding: 40px 120px;
     position: relative;
-    width: 1440px;
+    width: 100%;
   }
   
   .desktop .frame-14 {
