+++ client/src/style.css
@@ -1,0 +1,10 @@
+  body {
+    display: flex;
+    justify-content: center;
+    vertical-align: baseline;
+    min-height: 100vh;
+    background: #f0f0f0;
+  }
+  
+  
+
